Added OFPAT_EXP_WRITE_CONTEXT_TO_FIELD

+++ oflib/ofl-print.c
@@ -292,7 +292,7 @@ ofl_action_type_print(FILE *stream, uint16_t type)
 }
 
 char *
-ofl_oxm_type_to_string(uint16_t type)
+ofl_oxm_type_to_string(uint32_t type)
 {
     char *str;
     size_t str_size;
@@ -347,7 +347,7 @@ ofl_oxm_type_print(FILE *stream, uint32_t type)
     case OXM_OF_PBB_ISID:           {fprintf(stream, "pbb_isid"); return; }
     case OXM_OF_TUNNEL_ID:          {fprintf(stream, "tunnel_id"); return; }
     case OXM_OF_IPV6_EXTHDR:        {fprintf(stream, "ipv6_exthdr"); return; }
-    default: {                       fprintf(stream, "?(%d)", type); return; }    
+    default: {                       fprintf(stream, "?(%"PRIu32")", type); return; }    
     }
